refactor: Modal 컴포넌트 Base, Header, Content 분리
